Postgres to Datawarehouse Operator Implementation
This commit contains the implementation of PostgresToDwOperator
which is basically responsible for creating a uploading facts
and dimensional tables in the Postgres DW.

This Operator also interacts with a MySQL database for the data
that it will process. Lastly, a data model definition in a .sql file
was also added that is used by docker-compose to initialize
a star schema.